Improve circular dependencies
